requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic[email]==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0
python-dotenv==1.0.1
bcrypt==4.2.0
PyJWT==2.10.1
httpx==0.27.2
python-multipart==0.0.12

